standards pass: workspace boundary, decomposition, layout
- Promote target-projects/ to @goose-hub/target-projects workspace package
  and rewrite cli/server cross-package imports per FACTORY_RULES rule 28a.
  Expose targetProjectsRoot + skillsRoot for runtime FS lookups.
- Decompose oversized files: TimelineSection (1090->102),
  issues/service.ts (622->131, split into transitions/diff/triage/
  fake-run/internal), InboxList (533->105), InvestigationSection,
  QASection, CodeDiffSection, TaskHeader.
- Move IssueDiffDto to lib/types.ts; promote COST_PLACEHOLDER to
  lib/constants.ts; replace inline JSON parse in events/router with
  parseBody.
- Restructure inbox/ to follow components/+lib/+slice.test.ts pattern;
  promote CaptureModal to components/ui/ to drop the chrome->inbox
  cross-feature import; rename board/issue-card.test.ts to slice.test.ts.
- Update apps/web/STANDARDS.md to match the actual src/components/<feature>/
  layout (incl. roster) and add ADR 0015 for the workspace promotion.
