+++ apps/server/src/domains/issues/service.ts
@@ -1,62 +1,17 @@
-import { execFileSync } from 'node:child_process';
-import { existsSync, readFileSync } from 'node:fs';
-import { homedir } from 'node:os';
-import { join } from 'node:path';
-import { db } from '@goose-hub/core/db/db.js';
-import { events } from '@goose-hub/core/db/schema.js';
-import { type AgentEvent, eventStore } from '@goose-hub/core/event-stream/store.js';
-import { STATES } from '@goose-hub/core/state-machine/states.js';
-import type { StateName } from '@goose-hub/core/state-machine/states.js';
-import { isLegalTransition, legalTargets } from '@goose-hub/core/state-machine/transitions.js';
-import { cleanupWorktree } from '@goose-hub/core/workspaces/worktree.js';
-import { and, desc, eq, isNotNull } from 'drizzle-orm';
+import { eventStore } from '@goose-hub/core/event-stream/store.js';
 import { CACHE_KEY, bustCache, getCached } from '../../shared/cache.js';
 import type { Result } from '../../shared/middleware.js';
-import { getProject } from '../../shared/projects.js';
-import { getSourceForSlug, isValidSlug } from '../../shared/source.js';
+import { getSourceForSlug } from '../../shared/source.js';
+import { getLastPersonaIdsByWorkItem, getRepoRef } from './internal.js';
 
-// File is at apps/server/src/domains/issues/service.ts
-// 5 levels up from service.ts to apps/, then 1 more to repo root = 6 levels from the FILE
-// Using import.meta.dirname (directory): 5 levels up to repo root
-const REPO_ROOT = join(import.meta.dirname, '../../../../..');
-
-const OUTPUT_FIXTURES: Record<string, unknown> = {
-  triage: {
-    triage: { type: 'feature', priority: 'high' },
-    repoMatch: {
-      candidates: [
-        { repo: 'shaunnez/goose-hub', confidence: 87, evidence: 'keyword match', tier: 1 },
-      ],
-    },
-  },
-  investigate: {
-    findings: 'Root cause identified',
-    confidence: 'high',
-    recommendation: 'fix in core',
-  },
-};
-
-async function getRepoRef(slug: string): Promise<string> {
-  const cfg = await getProject(slug);
-  return cfg?.source.kind === 'github' ? cfg.source.repo : slug;
-}
-
-function getLastPersonaIdsByWorkItem(projectId: string): Map<string, string> {
-  const rows = db
-    .select({ workItemId: events.workItemId, personaId: events.personaId })
-    .from(events)
-    .where(and(eq(events.projectId, projectId), isNotNull(events.personaId)))
-    .orderBy(desc(events.id))
-    .all();
-
-  const map = new Map<string, string>();
-  for (const row of rows) {
-    if (row.workItemId != null && row.personaId != null && !map.has(row.workItemId)) {
-      map.set(row.workItemId, row.personaId);
-    }
-  }
-  return map;
-}
+// Public surface for the issues domain. The implementation is split across
+// sibling files to keep each concern focused; this barrel re-exports the
+// pieces the router and tests depend on.
+export { approveIssue, rejectIssue, transitionIssue } from './transitions.js';
+export type { TransitionResult } from './transitions.js';
+export { getIssueWorktreeDiff } from './diff.js';
+export { getIssueTriage, overrideIssueRepo } from './triage.js';
+export { fakeRun } from './fake-run.js';
 
 export async function listIssues(slug: string): Promise<Result<{ items: unknown[] }>> {
   const source = await getSourceForSlug(slug);
@@ -90,243 +45,6 @@ export async function getIssueEvents(
   return { ok: true, data: { events } };
 }
 
-/**
- * Approve gate (#186). Looks up the most recent `pr.opened` event for the
- * issue, merges that PR via the GitHub connector, emits `gate.approved` +
- * `pr.merged` events, and transitions factory:approved → factory:done.
- *
- * Optional `mergePRImpl` is dependency-injected so tests can stub the
- * REST call.
- */
-export async function approveIssue(
-  slug: string,
-  id: string,
-  options: {
-    mergePRImpl?: typeof import('@goose-hub/core/connectors/github/merge-pr.js').mergePR;
-  } = {},
-): Promise<Result<{ ok: true; sha: string; prNumber: number }>> {
-  const source = await getSourceForSlug(slug);
-  if (source == null) return { ok: false, error: 'project not found', status: 404 };
-  const repoRef = await getRepoRef(slug);
-  const workItemId = `github:${repoRef}#${id}`;
-
-  // Find the most recent pr.opened event for this issue.
-  const events = eventStore.replay({ projectId: slug, workItemId });
-  const prEvent = [...events].reverse().find((e) => e.kind === 'pr.opened');
-  if (prEvent == null) {
-    return {
-      ok: false,
-      error: 'no pr.opened event found — cannot approve without a PR',
-      status: 400,
-    };
-  }
-  const prNumber = (prEvent.payload as { prNumber?: number }).prNumber;
-  if (typeof prNumber !== 'number') {
-    return { ok: false, error: 'pr.opened event missing prNumber', status: 500 };
-  }
-
-  const token = process.env.GITHUB_TOKEN ?? '';
-  if (token.length === 0) {
-    return { ok: false, error: 'GITHUB_TOKEN env var not set', status: 500 };
-  }
-
-  const mergePR =
-    options.mergePRImpl ?? (await import('@goose-hub/core/connectors/github/merge-pr.js')).mergePR;
-
-  const merged = await mergePR({ repo: repoRef, prNumber, token });
-
-  eventStore.appendEvent({
-    projectId: slug,
-    workItemId,
-    kind: 'gate.approved',
-    payload: { source: 'ui', prNumber },
-  });
-  eventStore.appendEvent({
-    projectId: slug,
-    workItemId,
-    kind: 'pr.merged',
-    payload: { prNumber, sha: merged.sha },
-  });
-
-  // Clean up the dev worktree now that the PR is merged.
-  const allEvents = eventStore.replay({ workItemId });
-  const prOpenedEvent = allEvents
-    .slice()
-    .reverse()
-    .find((e) => e.kind === 'pr.opened');
-  if (prOpenedEvent != null) {
-    const { devRunId } = prOpenedEvent.payload as { devRunId?: string };
-    if (typeof devRunId === 'string') cleanupWorktree(devRunId);
-  }
-
-  await source.transitionState(id, 'factory:approved', 'factory:done');
-  await source.comment(id, `Approved via Goose Hub UI; PR #${prNumber} merged (${merged.sha}).`);
-
-  return { ok: true, data: { ok: true, sha: merged.sha, prNumber } };
-}
-
-/**
- * Reject gate (#186). Records a rejection note as a GitHub comment, emits
- * `gate.rejected`, and transitions factory:approved → factory:needs-fix.
- * Does NOT touch the PR (humans handle PR closure manually if needed).
- */
-export async function rejectIssue(
-  slug: string,
-  id: string,
-  reason: unknown,
-): Promise<Result<{ ok: true }>> {
-  if (typeof reason !== 'string' || reason.trim().length === 0) {
-    return { ok: false, error: 'rejection reason is required', status: 400 };
-  }
-  const source = await getSourceForSlug(slug);
-  if (source == null) return { ok: false, error: 'project not found', status: 404 };
-  const repoRef = await getRepoRef(slug);
-  const workItemId = `github:${repoRef}#${id}`;
-
-  eventStore.appendEvent({
-    projectId: slug,
-    workItemId,
-    kind: 'gate.rejected',
-    payload: { source: 'ui', reason },
-  });
-  await source.comment(id, `Rejected at approval gate: ${reason}`);
-  await source.transitionState(id, 'factory:approved', 'factory:needs-fix');
-
-  return { ok: true, data: { ok: true } };
-}
-
-/**
- * Live diff for the Code tab (#185). Returns the unified diff of the
- * current worktree for the most recent in-flight run on this issue, or
- * `{ diff: null }` when no worktree exists.
- *
- * Active-runId resolution: pick the latest event whose payload carries
- * a runId — preferring `pr.opened`, then `agent.implement-complete`,
- * then `agent.run-started` — and look for a worktree at
- * `~/.factory/workspaces/<runId>/`. If absent, return `{ diff: null }`.
- *
- * Trade-off: this is a one-shot diff, not a true SSE stream. The UI
- * polls (5 s default) — adequate for M2-style live UX. SSE-streamed
- * diff is M11+ work.
- */
-export async function getIssueWorktreeDiff(
-  slug: string,
-  id: string,
-  opts?: { fetchImpl?: typeof fetch },
-): Promise<Result<{ diff: string | null; runId: string | null; reason?: string }>> {
-  const fetchImpl = opts?.fetchImpl ?? fetch;
-  const source = await getSourceForSlug(slug);
-  if (source == null) return { ok: false, error: 'project not found', status: 404 };
-  if (!isValidSlug(slug)) return { ok: false, error: 'invalid slug', status: 400 };
-
-  const repoRef = await getRepoRef(slug);
-  const workItemId = `github:${repoRef}#${id}`;
-  const ascending = eventStore.replay({ projectId: slug, workItemId });
-
-  // Newest-first scan for a runId on a fix-issue lifecycle event.
-  const lifecycleKinds = new Set(['pr.opened', 'agent.implement-complete', 'agent.run-started']);
-  let runId: string | null = null;
-  for (let i = ascending.length - 1; i >= 0; i -= 1) {
-    const e = ascending[i];
-    if (lifecycleKinds.has(e.kind) && typeof e.runId === 'string' && e.runId.length > 0) {
-      runId = e.runId;
-      break;
-    }
-  }
-
-  if (runId == null) {
-    return {
-      ok: true,
-      data: { diff: null, runId: null, reason: 'no in-flight run for this issue' },
-    };
-  }
-
-  const worktreePath = join(homedir(), '.factory', 'workspaces', runId);
-  if (!existsSync(worktreePath)) {
-    const prResult = await tryGitHubPrDiff(ascending, repoRef, fetchImpl);
-    if (prResult != null)
-      return { ok: true, data: { diff: prResult.diff, runId: prResult.prRunId } };
-    return {
-      ok: true,
-      data: { diff: null, runId, reason: 'worktree not found (cleaned up or pre-creation)' },
-    };
-  }
-
-  if (!existsSync(join(worktreePath, '.git'))) {
-    const prResult = await tryGitHubPrDiff(ascending, repoRef, fetchImpl);
-    if (prResult != null)
-      return { ok: true, data: { diff: prResult.diff, runId: prResult.prRunId } };
-    return {
-      ok: true,
-      data: {
-        diff: null,
-        runId,
-        reason: 'worktree directory exists but git repo not yet initialised',
-      },
-    };
-  }
-
-  try {
-    const diff = execFileSync('git', ['diff', 'HEAD', '--', '.', ':(exclude).claude/'], {
-      cwd: worktreePath,
-      encoding: 'utf8',
-      maxBuffer: 4 * 1024 * 1024,
-    });
-    if (diff.length > 0) return { ok: true, data: { diff, runId } };
-    // Worktree exists but all changes are committed (e.g. PR already opened).
-    // Fall through to the GitHub PR diff.
-    const prResult = await tryGitHubPrDiff(ascending, repoRef, fetchImpl);
-    if (prResult != null)
-      return { ok: true, data: { diff: prResult.diff, runId: prResult.prRunId } };
-    return {
-      ok: true,
-      data: { diff: null, runId, reason: 'no uncommitted changes and no PR diff available' },
-    };
-  } catch (err) {
-    return {
-      ok: true,
-      data: {
-        diff: null,
-        runId,
-        reason: `git diff failed: ${(err as Error).message}`,
-      },
-    };
-  }
-}
-
-async function tryGitHubPrDiff(
-  ascending: AgentEvent[],
-  repo: string,
-  fetchImpl: typeof fetch,
-): Promise<{ diff: string; prRunId: string | null } | null> {
-  const token = process.env.GITHUB_TOKEN ?? '';
-  if (token.length === 0) return null;
-  const prEvent = [...ascending]
-    .reverse()
-    .find(
-      (e) =>
-        e.kind === 'pr.opened' &&
-        typeof (e.payload as Record<string, unknown>)?.prNumber === 'number',
-    );
-  if (prEvent == null) return null;
-  const prNumber = (prEvent.payload as { prNumber: number }).prNumber;
-  const prRunId = typeof prEvent.runId === 'string' ? prEvent.runId : null;
-  try {
-    const res = await fetchImpl(`https://api.github.com/repos/${repo}/pulls/${prNumber}`, {
-      headers: {
-        Authorization: `Bearer ${token}`,
-        Accept: 'application/vnd.github.v3.diff',
-        'X-GitHub-Api-Version': '2022-11-28',
-      },
-    });
-    if (!res.ok) return null;
-    const diff = await res.text();
-    return { diff, prRunId };
-  } catch {
-    return null;
-  }
-}
-
 export async function getIssueComments(
   slug: string,
   id: string,
@@ -337,109 +55,6 @@ export async function getIssueComments(
   const workItemId = `github:${repoRef}#${id}`;
   const comments = await source.listComments(workItemId);
   return { ok: true, data: { comments } };
-}
-
-interface TriageEventPayload {
-  triage: { type: string; priority: string };
-  repoMatch: {
-    candidates: Array<{ repo: string; confidence: number; evidence: string; tier: number }>;
-  };
-}
-
-interface TriageDto {
-  type: string;
-  priority: string;
-  candidates: Array<{ repo: string; confidence: number; evidence: string; tier: number }>;
-  overrideRepo: string | null;
-}
-
-/**
- * Build a TriageDto from a triage-complete event payload (#204).
- * Both getIssueTriage and overrideIssueRepo construct identical-shaped
- * triage payloads from the latest agent.triage-complete event; this helper
- * keeps them in sync as the triage schema evolves.
- */
-function buildTriageDto(payload: unknown, overrideRepo: string | null): TriageDto {
-  const p = payload as TriageEventPayload;
-  return {
-    type: p.triage.type,
-    priority: p.triage.priority,
-    candidates: p.repoMatch.candidates ?? [],
-    overrideRepo,
-  };
-}
-
-export async function getIssueTriage(
-  slug: string,
-  id: string,
-): Promise<Result<{ triage: unknown }>> {
-  const source = await getSourceForSlug(slug);
-  if (source == null) return { ok: false, error: 'project not found', status: 404 };
-  const workItemId = `github:${source.repoRef}#${id}`;
-  const projectId = source.projectId;
-
-  const allEvents = eventStore.replay({ projectId, workItemId });
-  const triageEvent = allEvents.filter((e) => e.kind === 'agent.triage-complete').at(-1);
-  if (triageEvent == null) return { ok: true, data: { triage: null } };
-
-  const overrideEvent = allEvents.filter((e) => e.kind === 'agent.repo-override').at(-1);
-  const overridePayload = overrideEvent?.payload as { repo?: string } | undefined;
-
-  return {
-    ok: true,
-    data: {
-      triage: buildTriageDto(triageEvent.payload, overridePayload?.repo ?? null),
-    },
-  };
-}
-
-export type TransitionResult =
-  | { ok: true; data: { ok: true; from: StateName; to: StateName } }
-  | { ok: false; error: string; status: number; legalTargets?: readonly StateName[] };
-
-export async function transitionIssue(
-  slug: string,
-  id: string,
-  from: unknown,
-  to: unknown,
-): Promise<TransitionResult> {
-  if (from == null || to == null) {
-    return { ok: false, error: "missing 'from' or 'to'", status: 400 };
-  }
-  if (!(STATES as readonly string[]).includes(from as string)) {
-    return { ok: false, error: `invalid state name for 'from': ${from}`, status: 400 };
-  }
-  if (!(STATES as readonly string[]).includes(to as string)) {
-    return { ok: false, error: `invalid state name for 'to': ${to}`, status: 400 };
-  }
-
-  const fromState = from as StateName;
-  const toState = to as StateName;
-
-  if (!isLegalTransition(fromState, toState)) {
-    return {
-      ok: false,
-      error: 'illegal transition',
-      status: 422,
-      legalTargets: legalTargets(fromState),
-    };
-  }
-
-  const source = await getSourceForSlug(slug);
-  if (source == null) return { ok: false, error: 'project not found', status: 404 };
-
-  const workItemId = `github:${source.repoRef}#${id}`;
-  await source.transitionState(workItemId, fromState, toState);
-
-  eventStore.appendEvent({
-    projectId: slug,
-    workItemId,
-    kind: 'state.transitioned',
-    payload: { from: fromState, to: toState, by: 'ui' },
-  });
-
-  bustCache(CACHE_KEY.issues(slug));
-  return { ok: true, data: { ok: true, from: fromState, to: toState } };
 }
 
 export async function commentOnIssue(
@@ -513,110 +128,4 @@ export async function setIssueLabel(
   });
   bustCache(CACHE_KEY.issues(slug));
   return { ok: true, data: { ok: true } };
-}
-
-export async function overrideIssueRepo(
-  slug: string,
-  id: string,
-  repo: unknown,
-): Promise<Result<{ triage: unknown }>> {
-  if (typeof repo !== 'string') return { ok: false, error: 'repo is required', status: 400 };
-  if (!isValidSlug(slug)) return { ok: false, error: 'invalid slug', status: 400 };
-
-  const source = await getSourceForSlug(slug);
-  if (source == null) return { ok: false, error: 'project not found', status: 404 };
-
-  const reposMdPath = join(REPO_ROOT, 'target-projects', slug, 'repos.md');
-  const reposMd = readFileSync(reposMdPath, 'utf8');
-  const allowedRepos =
-    reposMd
-      .match(/^###\s+\[([^\]]+)\]/gm)
-      ?.map((m) => m.replace(/^###\s+\[/, '').replace(/\]$/, '')) ?? [];
-
-  if (!allowedRepos.includes(repo)) {
-    return { ok: false, error: `repo '${repo}' not in allowlist`, status: 400 };
-  }
-
-  const workItemId = `github:${source.repoRef}#${id}`;
-  const projectId = source.projectId;
-
-  eventStore.appendEvent({ projectId, workItemId, kind: 'agent.repo-override', payload: { repo } });
-
-  const allEvents = eventStore.replay({ projectId, workItemId });
-  const triageEvent = allEvents.filter((e) => e.kind === 'agent.triage-complete').at(-1);
-  if (triageEvent == null) return { ok: true, data: { triage: null } };
-
-  return {
-    ok: true,
-    data: {
-      triage: buildTriageDto(triageEvent.payload, repo),
-    },
-  };
-}
-
-export async function fakeRun(
-  slug: string,
-  id: string,
-  skill: string,
-): Promise<Result<{ ok: true; skill: string }>> {
-  // NODE_ENV guard (#203). fakeRun emits synthetic agent.* events that
-  // pollute the durable SQLite event log. Disable in production so a
-  // misrouted call cannot corrupt timeline debugging.
-  if (process.env.NODE_ENV === 'production') {
-    return { ok: false, error: 'fake-run is disabled in production', status: 404 };
-  }
-
-  const safeSkill = skill === 'investigate' ? 'investigate' : 'triage';
-
-  const source = await getSourceForSlug(slug);
-  if (source == null) return { ok: false, error: 'project not found', status: 404 };
-
-  const repoRef = await getRepoRef(slug);
-  const workItemId = `github:${repoRef}#${id}`;
-
-  const LOG_LINES = [
-    'Fetching issue metadata from GitHub...',
-    'Parsing labels and body content...',
-    'Scoring priority and work type...',
-    'Drafting decision summary...',
-    'Finalising structured output...',
-  ];
-
-  (async () => {
-    eventStore.appendEvent({
-      projectId: slug,
-      workItemId,
-      kind: 'agent.spawned',
-      payload: { skill: safeSkill },
-    });
-    await new Promise((r) => setTimeout(r, 700));
-    for (const line of LOG_LINES) {
-      eventStore.appendEvent({ projectId: slug, workItemId, kind: 'agent.log', payload: { line } });
-      await new Promise((r) => setTimeout(r, 600));
-    }
-    eventStore.appendEvent({
-      projectId: slug,
-      workItemId,
-      kind: 'agent.decision-summary',
-      payload: { summary: `Running ${safeSkill} skill on issue #${id}` },
-    });
-    await new Promise((r) => setTimeout(r, 700));
-    if (safeSkill === 'triage') {
-      const fixture = OUTPUT_FIXTURES.triage as { triage: unknown; repoMatch: unknown };
-      eventStore.appendEvent({
-        projectId: slug,
-        workItemId,
-        kind: 'agent.triage-complete',
-        payload: fixture,
-      });
-    }
-    eventStore.appendEvent({
-      projectId: slug,
-      workItemId,
-      kind: 'agent.terminated',
-      payload: { skill: safeSkill, status: 'completed', output: OUTPUT_FIXTURES[safeSkill] },
-    });
-  })();
-
-  return { ok: true, data: { ok: true, skill: safeSkill } };
 }
