+++ pnpm-workspace.yaml
@@ -2,3 +2,4 @@ packages:
   - 'apps/*'
   - 'core'
   - 'skills'
+  - 'target-projects'
