+++ target-projects/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@goose-hub/skills",
+  "name": "@goose-hub/target-projects",
   "version": "0.0.1",
   "private": true,
   "type": "module",
